move tests to lib.rs
