+++ rust/sdk/src/builder/processor_builder.rs
@@ -4,9 +4,7 @@ use crate::{
     types::transaction_context::TransactionContext,
 };
 use anyhow::Result;
-use instrumented_channel::{
-    instrumented_bounded_channel, InstrumentedAsyncReceiver, InstrumentedAsyncSender,
-};
+use instrumented_channel::{instrumented_bounded_channel, InstrumentedAsyncReceiver};
 use petgraph::{
     dot::Config,
     graph::{DiGraph, EdgeReference, NodeIndex},
