+++ rust/sdk/src/steps/transaction_stream_step.rs
@@ -134,9 +134,8 @@ mod tests {
     use crate::{
         builder::ProcessorBuilder,
         test::{steps::pass_through_step::PassThroughStep, utils::receive_with_timeout},
-        traits::{IntoRunnableStep, RunnableStepWithInputReceiver},
+        traits::IntoRunnableStep,
     };
-    use instrumented_channel::instrumented_bounded_channel;
     use std::time::Duration;
 
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
